Move some properties into utils
